assignment-5/pyproject.toml
[tool.poetry]
name = "ds561-a5"
version = "0.1.0"
description = ""
authors = ["Alicja Mahr <bmahr@bu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.0"
google-cloud-pubsub = "^2.18.4"
google-cloud-storage = "^2.12.0"
google-cloud-logging = "^3.8.0"
waitress = "^2.1.2"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.22"
cloud-sql-python-connector = {extras = ["pymysql"], version = "^1.4.2"}

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
